% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.R
\name{varMatInv}
\alias{aee}
\alias{varMatInv}
\title{Generate an inversed variance matrix from expression profile across species}
\usage{
varMatInv(objects, phy, taxa = "all", subtaxa)

aee(x, phy, mat, select = c("all", "descendents"), CI = TRUE)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{phy}{an rooted expression character tree}

\item{taxa}{oen single character or a vector of characters sepcifying taxa to generate
an inversed variance matrix.
If one single character "all" is given,
all the taxa included in the \code{taxaExp} will be matched and included ("all" by default).}

\item{subtaxa}{one single character specifying sub taxa to be included in generating
an inversed variance matrix}

\item{x}{a vector of known expression profile, preferably log-transformed expression levels (e.g. log RPKM)}

\item{mat}{a matrix generated from "varMatInv" function}

\item{select}{indicate if descendents of the node or all tips should be used}

\item{CI}{a logical specifying whether to return the 95% confidence intervals
of the estimated ancestral expression levels}

\item{phy}{a phylogenetic tree in the form of object "phylo"}
}
\value{
returns an inversed variance matrix

returns a list containing estimated ancestral expression profile
as well as other requested parameters
}
\description{
This function generate an inversed variance matrix from expression profiles
of one-to-one orthologous genes across species

This function esitmates ancestral expression profile and related statistical uncertainty
}
\examples{

data('tetraexp')
dismat <- expdist(tetraexp.objects, taxa = "all", subtaxa = "Brain", method = "sou")
exp_tree <- NJ(dismat)
exp_tree <- root(exp_tree, outgroup = "Chicken_Brain", resolve.root = T)
exp_tree <- no0br(exp_tree)
var_mat <- varMatInv(objects = tetraexp.objects,phy = exp_tree,taxa = "all", subtaxa = "Brain")
exp_table <- exptabTE(tetraexp.objects, taxa = "all", subtaxa = "Brain")
exp_one <- aee(exp_table[1,], exp_tree, var_mat)
exp_tree$node.label <- exp_one$est
plot(exp_tree)

}

