% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot.exphy}
\alias{boot.exphy}
\title{Bootstrapping expression phylogeny}
\usage{
boot.exphy(phy = NULL, objects = NULL, rowindex = NULL,
  method = c("sou", "ced", "pea", "souln", "nbdln", "euc", "cos", "jsd"),
  B = 100, rooted = NULL, trees = FALSE)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}.}

\item{rowindex}{a vector of numbers corresponded to indices of selecting rows}

\item{method}{specifying which distance method to be used
to estimate expression phylogeny in bootstrapping.}

\item{B}{the number of bootstrap replicates.}

\item{rooted}{if "phy" is a rooted tree, a character of the root node's label when constructing "phy";
if "phy" is unrooted tree, NULL (NULL by default).}

\item{trees}{a logical specifying whether to return the bootstrapped trees (FALSE by default).}
}
\value{
similar to boot.phylo in ape, boot.exptree returns a numeric vector
which ith element is the number associated to the ith node of phy.
if trees = TRUE, boot.exptree returns a list whose first element (named "BP") is like before,
and the second element("trees") is a list with the bootstrapped trees.
}
\description{
bootstrap by resampling gene (gene, transcript, exon, etc..)
}
\examples{

data(tetraexp)
dismat <- expdist(tetraexp.objects, taxa = "all",
                 subtaxa = "Brain",
                 method = "sou")
tr <- root(NJ(dismat), "Chicken_Brain")
plot(tr)
bs <- boot.exphy(tr, tetraexp.objects, method = "sou",
                 B = 100, rooted = "Chicken_Brain")
nodelabels(bs)

}
\author{
Hang Ruan (hang.ruan@hotmail.com).
}

