% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{estdelta}
\alias{estdelta}
\title{Estimating delta values from a distance matrix.}
\usage{
estdelta(disMat = NULL)
}
\arguments{
\item{disMat}{a distance matrix with column and row names}
}
\value{
returns a list of two,
quartet: vector of delta numbers,
each corresponded to a quartet in the distance matrix.
taxa: matrix of delta numbers,
colnames corresponded to taxaNames
taxaNames: vector of taxaNames
}
\details{

}
\examples{
data(tetraexp)
dis.mat <- expdist(tetraexp.objects, taxa = "all",
                     subtaxa = "Brain",
                     method = "pea")
deltas <- estdelta(dis.mat)
hist(deltas$quartet)

}
\references{
Holland,B.R. et al. 2002. Delta plots: a tool for analyzing phylogenetic distance data.
Mol. Biol. Evol., 19, 2051–2059.
}

