% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_Vmu.R
\name{dist.sou_v}
\alias{dist.sou_v}
\title{Estimate sOU expression distance when expression optima vary among genes}
\usage{
dist.sou_v(exptable = NULL)
}
\arguments{
\item{exptable}{an exprssion level table: column corresponds to median
expression value of all biological samples within one taxa_subtaxa group;
row corresponds to othologous genes
('\code{spe}' for Spearman's correlation coefficient; \code{pea}' for
Pearson's correlation correlation coefficient).}
}
\value{
returns a list containing estimated expression distance
(sOU_v) matrix and parameter \code{pi} which measures the variance
 of optima among genes under sOU model
}
\description{
estimate sOU expression distance when expression optima vary among
 genes (sOU_v distance)
}
\examples{
data('tetraExp')
exp_table <- exptabTE(tetraExp, taxa = "all", subtaxa = "heart")
dismat <- dist.sou_v(exptable = exp_table)
dismat
}
\author{
Jingwen Yang
}
