% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{dist.brownian}
\alias{dist.cos}
\alias{dist.euc}
\alias{dist.gu2013}
\alias{dist.jsd}
\alias{dist.rho}
\alias{distances}
\title{Internal functions for estimating pair-wise expression distances}
\usage{
dist.gu2013(reads.count = NULL, gene_length = NULL, omega = NULL)

dist.brownian(reads.count = NULL, gene_length = NULL)

dist.jsd(meanRPKM = NULL, taxon.names = NULL)

dist.rho(meanRPKM = NULL)

dist.euc(meanRPKM = NULL)

dist.cos(meanRPKM = NULL)
}
\description{
Several published methods to estimate pair-wise expression distances
}
\references{

}

