% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{estomega}
\alias{estomega}
\title{Estimation of over-dispersion parameter omega}
\usage{
estomega(objects = NULL, overwrite = TRUE)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{overwrite}{a logical specifying whether to overwrite the existing omega value
stored in the \code{taxonExp} objects}
}
\examples{
data(tetraexp)
tetraexp.objects.with.omega <- estomega(tetraexp.objects)
tetraexp.objects.with.omega[[1]]$omega

}
\references{
Gu,X. et al. 2013. Phylogenomic distance method for analyzing transcriptome evolution based on RNA-seq data.
Genome Biol Evol, 5, 1746-1753.
}

