% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{estomega.sample}
\alias{estomega.sample}
\alias{.estomega.sample}
\title{Internal function for omega estimation in bootstrapping}
\usage{
.estomega.sample(objects = NULL, geneIndex = NULL)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{geneIndex}{a vector of numbers corresponded to indices of selecting rows}
}
\value{
NULL.
}
\description{
Internal function for omega estimation in bootstrapping
}
