% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation.R
\name{estParaGamma}
\alias{estParaGamma}
\title{Estimation of parameters of selection strength Gamma Distribution}
\usage{
estParaGamma(exptable = NULL, corrmatinv = NULL)
}
\arguments{
\item{exptable}{an expression level table: column corresponds to median expression value of all biological samples
within one taxa_subtaxa group; row corresponds to othologous genes}

\item{corrmatinv}{an inversed correlation matrix between specie}
}
\value{
returns a vector of parameters estimated from gamma distribution
}
\description{
Estimation of parameters of selection strength Gamma Distribution
}

