% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation.R
\name{estParaQ}
\alias{estParaQ}
\title{Estimation of parameter Q from expression table}
\usage{
estParaQ(exptable = NULL, corrmatinv = NULL)
}
\arguments{
\item{exptable}{an exprssion level table: column corresponds to median expression value of all biological samples
within one taxa_subtaxa group; row corresponds to othologous genes}

\item{corrmatinv}{an inversed correlation matrix between species}
}
\value{
returns a vector of parameter Qs from given expression table
}
\description{
Estimation of parameter Q from expression table
}

