% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{estVarCov}
\alias{estVarCov}
\title{Expression variance-covariance matrix}
\usage{
estVarCov(expMat = NULL)
}
\arguments{
\item{expMat}{an expression distance matrix}
}
\value{
returns an expression variance-covariance matrix
}
\description{
Generate a variance-covariance matrix from an expression distance matrix
}

