% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example of the Input Data Format}
\description{
A Table shows the format of the input data.
}
\references{
Xu,C. et al. (2018) Human-specific features of spatial gene expression and regulation
in eight brain regions. Genome Research, 28, 1097–1110.
}
