% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdist.R
\name{expdist}
\alias{expdist}
\title{Generate a distance matrix from a taxaExp class}
\usage{
expdist(objects = NULL, taxa = "all", subtaxa = "all", method = c("rho",
  "brownian", "gu2013", "euc", "cos", "jsd"))
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{taxa}{one single character or a vector of characters specifying main taxa selected for
calculating expression distance.
If one single character "all" is given,
all the taxa included in the \code{taxaExp} will be matched and selected ("all" by default).}

\item{subtaxa}{one single character or a vector of characters sepcifying sub taxa selected for
calculating expression distance.
If one singke character "all" is given,
all the subtaxa included in the \code{taxaExp} will be matched and selected ("all" by default).}

\item{method}{specifying which distance method to be used
to estimate expression phylogeny in bootstrapping.}
}
\value{
returns an expression distance matrix
}
\examples{
data(tetraexp)
dismat <- expdist(tetraexp.objects, taxa = "all",
                 subtaxa = "Brain",
                 method = "rho")
tr <- root(nj(dismat), "Chicken_Brain")
plot(tr)

}
\references{

}

