% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{exptabTE}
\alias{exptabTE}
\title{Expression level table from a taxaExp class}
\usage{
exptabTE(objects = NULL, taxa = "all", subtaxa = "all", rowindex = NULL,
  logrithm = TRUE)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{taxa}{one single character or a vector of characters specifying main taxa to be included in
the expression level table.
If one single character "all" is given,
all the taxa included in the \code{taxaExp} will be matched and included ("all" by default).}

\item{subtaxa}{one single character or a vector of characters specifying sub taxa to be included in
the expression level table.
If one single character "all" is given,
all the subtaxa included in the \code{taxaExp} will be matched and included ("all" by default).}

\item{rowindex}{a vector of numbers corresponded to indices of selecting rows}

\item{logrithm}{a logical specifying whether to apply expression value log2 tranformation (TRUE by default).}
}
\value{
an expression level table: column corresponds to median expression value of all biological samples
within one taxa_subtaxa group; row corresponds to othologous genes
}
\description{
Generate an expression level table from a taxaExp class
}

