% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{estgamma}
\alias{estgamma}
\title{Estimating gamma from taxaExp objects.}
\usage{
estgamma(objects = NULL, taxa = NULL, subtaxa = NULL, method = c("rho",
  "brownian", "gu2013", "euc", "cos", "jsd"))
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{taxa}{one single character or a vector of characters specifying main taxa selected for
calculating gammaD and gammaE.
If one single character "all" is given,
all the taxa included in the \code{taxaExp} will be matched and selected ("all" by default).}

\item{subtaxa}{one single character or a vector of characters sepcifying sub taxa selected for
calculating gammaD and gammaE.
#' If one single character "all" is given,
all the taxa included in the \code{taxaExp} will be matched and selected ("all" by default).}

\item{method}{specifying which distance method to be used
to estimate expression phylogeny in bootstrapping.}
}
\value{
returns a data frame of six columns, refer to details for more information.
}
\details{

}
\examples{
data(tetraexp)
gamma.df <- estgamma(tetraexp.objects, taxa = "all",
                     subtaxa = c("Brain", "Cerebellum"),
                     method = "rho")
gamma.df

}
\references{

}

