% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters-data.R
\docType{data}
\name{genefilters}
\alias{genefilters}
\title{Filters for subsetting orthologous genes}
\format{Two dataframes whose first column corresponds to Ensembl gene ids,
second column corresponds to housekeeping gene symbol,
and third column corresponds to Go ids related to Nervous System Development.}
\description{
Two example filters for subsetting orthologous genes. 
1. Housekeeping genes
2. Nervous System Development genes
}
\examples{
data(genefilters)
head(tetrafilter)

}
\references{
Brawand,D. et al. (2011) The evolution of gene expression levels in mammalian organs.
Nature, 478, 343-348.
Dorus S. et al. (2004) Accelerated evolution of nervous system genes in the origin of Homo sapiens. 
Cell 119:1027-1040.
Eisenberg E, Levanon EY. (2013) Human housekeeping genes, revisited. 
Trends Genet 29:569-574.
}

