% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map.ls}
\alias{map.ls}
\title{Map the expression distance onto a given tree topology}
\usage{
map.ls(phy, D)
}
\arguments{
\item{phy}{a phylogenetic tree}

\item{D}{a distance matrix has row&column names corresponded to tip labels}
}
\value{
mapped tree and the sum of squares residual error (Q-score)
}
\description{
This function computes the least squares brach lengths
on a tree with known topology and a distance matrix. Borrowed something from
"phytools" package.
}
\examples{
map.tree <- map.ls(phy, D)
attr(map.tree, "Q-score")

}

