% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nj.R
\name{NJ}
\alias{NJ}
\title{Neighbor-joining}
\usage{
NJ(x)
}
\description{
The famous neighbor-joining tree estimation function from Saitou and Nei (1987).
}
\examples{

data(tetraexp)
dismat <- expdist(tetraexp.objects, taxa = "all",
                 subtaxa = "Brain",
                 method = "pea")
tr <- root(NJ(dismat), "Chicken_Brain")
plot(tr)

}

