% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.R
\name{no0br}
\alias{no0br}
\title{No zero branch length}
\usage{
no0br(phy)
}
\arguments{
\item{phy}{an rooted tree}
}
\value{
returns the tree without zero-length or negative-length branch
}
\description{
This function does a small tweak on the tree to remove zero-length branch
between root and MRCA of the ingroup, and replace any negative branch lengths
with a small length
}

