% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primatexp-data.R
\docType{data}
\name{primatexp}
\alias{primatexp}
\title{taxaExp object created from expression data of primate species}
\format{a \code{taxaExp} object, contains 35 \code{taxonExp}:
with raw reads count data and TPM value on 13278 genes.}
\description{
large \code{taxaExp} object created from
six tissues' expression data of six primate species
}
\examples{
data(primatexp)
primatexp.objects
primatexp.objects[[1]]
}
\references{
Brawand,D. et al. (2011) The evolution of gene expression levels in mammalian organs.
Nature, 478, 343-348.
}

