% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.taxaExp}
\alias{print.taxonExp}
\title{Consice display of \code{taxaExp} or \code{taxonExp} object}
\usage{
print(objects, ...)

\method{print}{taxaExp}(objects, details = FALSE, ...)

\method{print}{taxonExp}(objects, printlen = 6, ...)
}
\arguments{
\item{objects}{an object of class \code{taxaExp} or class \code{taxonExp}.}

\item{...}{further arguments passed to or from other methods.}

\item{details}{a logical specifying whether to print taxa and subtaxa names.}

\item{printlen}{the number of biological replicates title to print (6 by default).}
}
\value{
returns a consice display of \code{taxaExp} or \code{taxonExp} object
}
\description{
Consice display of \code{taxaExp} or \code{taxonExp} object
}
\examples{
data(tetraExp)
print(tetraExp, details = TRUE)
print(tetraExp[[1]], printlen = 6)


}
