% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.taxaExp}
\alias{print.taxaExp}
\title{Consice display of a taxaExp object}
\usage{
\method{print}{taxaExp}(objects, details = FALSE, ...)
}
\arguments{
\item{objects}{an object of class \code{taxaExp}.}

\item{details}{a logical specifying whether to print taxa and subtaxa names.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
NULL.
}
\examples{
data(tetraexp)
print(tetraexp.objects, details = TRUE)

}

