% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.taxonExp}
\alias{print.taxonExp}
\title{Concise display of a taxonExp object}
\usage{
\method{print}{taxonExp}(object, printlen = 6, ...)
}
\arguments{
\item{object}{an object of class \code{taxonExp}.}

\item{printlen}{the number of biological replicates title to print (6 by default).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
NULL.
}
\description{
Concise display of a taxonExp object
}
\examples{
data(tetraexp)
print(tetraexp.objects[[1]], printlen = 6)

}

