% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeTest.R
\name{RelaRate.test}
\alias{RelaRate.test}
\title{Relative Rate Test for Gene Expression}
\usage{
RelaRate.test(expTable = NULL, x = NULL, y = NULL, outgroup = NULL,
  Pi = 0, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{expTable}{an exprssion level table: column corresponds to median
expression value of all biological samples within one taxa_subtaxa group;
row corresponds to othologous genes}

\item{x}{taxa name}

\item{y}{taxa name}

\item{outgroup}{outgroup name}

\item{Pi}{value of parameter Pi which measures the variance of optima
among genes under sOU model. Pi = 0 by default.}

\item{alternative}{character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}
}
\value{
A list containing the following components:
\itemize{
\item Z_score the z-statistic
\item alternative    records the value of the input argument alternative: "greater", "less" or "two.sided".
\item p.value    the p-value for the test
}
}
\description{
statistically compare the relative expression rate
of a given set of genes between two taxa (or species)
}
\examples{
data(tetraExp)
exp_table <-exptabTE(tetraExp, taxa = 'all', subtaxa = 'Brain',rowindex = 1:100)
ztest <- RelaRate.test(expTable = exp_table, x = 'human', y = 'chimpanzee',
outgroup = 'macaque', alternative = 'greater')
ztest
}
\author{
Jingwen Yang
}
