% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{TEconstruct}
\alias{TEconstruct}
\title{Construct a taxaExp object.}
\usage{
TEconstruct(readsCountFP = NULL, geneInfoFP = NULL, taxa = "all",
  subtaxa = "all", calRPKM = TRUE, rmOut = TRUE, verbose = FALSE)
}
\arguments{
\item{readsCountFP}{a text file contains raw reads count data.
Row names correspond with gene names,
and column names correspond with taxon and subtaxon names.}

\item{geneInfoFP}{a text file contains gene length information.
Row names conrrespond with gene names,
and column names correspond with taxon names.}

\item{taxa}{one single character or a vector of characters specifying main taxa selected for
constructing \code{taxaExp} object.
Taxa names are extracted from row names given in gene length file.
If one single character "all" is given,
all the taxa in the row names will be matched and selected ("all" by default).}

\item{subtaxa}{one single character or a vector of characters sepcifying sub taxa selected for
constructing \code{taxaExp} object.
If one single character "all" is given,
all the sub taxa in the row names will be matched and selected ("all" by default).}

\item{calRPKM}{a logical sepcifying whether to calculate RPKM value
while constructing \code{taxaExp} object (TRUE by default).}

\item{rmOut}{a logical sepcifying whether to remove reads count outliers
while constructing \code{taxaExp} objects (TRUE by default).}

\item{verbose}{a logical specifying whether to print more information on the screen
while constructing \code{taxaExp} objects (FALSE by default).}
}
\value{
returns an object of class \code{Taxa} (S3 class, a list of \code{taxonExp} objects).
}
\description{
constructor function for \code{taxaExp} objects
This function takes in a reads count file and a gene information file.
And construct a \code{taxaExp} object from which user can extract information
for display or for further analysis.
}
\examples{

taxa.objects = TEconstruct(readsCountFP = system.file('extdata/tetraexp.reads.count.raw.txt', package = 'phyExp'),
   geneInfoFP = system.file('extdata/tetraexp.gene.length.ortholog.txt', package = 'phyExp'),
   taxa = "all", subtaxa = c("Brain", "Cerebellum"), calRPKM = TRUE, rmOut =TRUE)

}

