% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetraexp-data.R
\docType{data}
\name{tetraexp}
\alias{tetraexp}
\title{taxaExp object created from expression data}
\format{a \code{taxaExp} object, contains 53 \code{taxonExp}:
with raw reads count data and TPM value on 5636 genes.}
\description{
large \code{taxaExp} object created from
six tissues' expression data of nine tetrapod species
}
\examples{
data(tetraexp)
tetraexp.objects
tetraexp.objects[[1]]
}
\references{
Brawand,D. et al. (2011) The evolution of gene expression levels in mammalian organs.
Nature, 478, 343-348.
}

