% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{esttheta}
\alias{esttheta}
\title{Estimating theta values from a distance matrix.}
\usage{
esttheta(disMat = NULL)
}
\arguments{
\item{disMat}{a distance matrix with column and row names}
}
\value{
returns a list of two,
quartet: vector of theta numbers,
each corresponded to a quartet in the distance matrix.
taxa: matrix of theta numbers,
colnames corresponded to taxaNames
taxaNames: vector of taxaNames
}
\details{

}
\examples{
data(tetraexp)
dis.mat <- expdist(tetraexp.objects, taxa = "all",
                     subtaxa = "Brain",
                     method = "pea")
thetas <- esttheta(dis.mat)
hist(thetas$quartet)

}
\references{

}

