% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeExp-package.R
\docType{package}
\name{TreeExp-package}
\alias{TreeExp-package}
\title{the TreeExp package for phylogenetic analysis of expression data}
\description{
\code{TreeExp} is an R package that performs expression phylogenetic analysis
from RNA-seq count data, including optimized input formatting, normalization
and pair-wise distance evaluation, preliminary phylogenetic network analysis.
}
\author{
Hang Ruan (hang.ruan@hotmail.com).
}

