% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees-data.R
\docType{data}
\name{trees}
\alias{trees}
\title{species trees and ultrametric time trees}
\format{several trees in the format \code{phylo}}
\description{
species trees and ultrametric time trees of 
nine tetrapod species as well as six primate species
}
\examples{
data(trees)
plot(tetraspeciestree)
plot(tetratimetree)
plot(primatespeciestree)
plot(primatetimetree)
}
\references{
\link{http://www.ensembl.org/info/about/speciestree.html}
\link{http://www.timetree.org}
}

