% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.R
\name{varMatInv}
\alias{varMatInv}
\title{Generate an inversed variance matrix from expression profile
across species}
\usage{
varMatInv(objects, phy, taxa = "all", subtaxa)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or
an object of class \code{taxaExp}}

\item{phy}{an expression character tree}

\item{taxa}{one single character or a vector of characters sepcifying
taxa to generate an inversed variance matrix. If one single character
"all" is given, all the taxa included in the \code{taxaExp}
will be matched and included ("all" by default).}

\item{subtaxa}{one single character specifying sub taxa to be
included in generating an inversed variance matrix}
}
\value{
returns an inversed variance matrix
}
\description{
This function generate an inversed variance matrix
from expression profiles of one-to-one orthologous genes across species
}
