% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.R
\name{aee}
\alias{aee}
\title{Ancestral Expression Estimation}
\usage{
aee(x, phy, mat, CI = TRUE)
}
\arguments{
\item{x}{a vector of known expression profile, preferably log-transformed
expression levels (e.g. log RPKM)}

\item{phy}{an unrooted phylogenetic tree in the form of class "phylo"}

\item{mat}{a matrix generated from "varMatInv" function}

\item{CI}{a logical specifying whether to return the 95% confidence
intervals of the estimated ancestral expression levels}
}
\value{
returns a list containing estimated ancestral expression profile
as well as other requested parameters
}
\description{
This function esitmates ancestral expression profile and
related statistical uncertainty
}
\examples{
library('ape')
data('tetraExp')
dismat <- expdist(tetraExp, taxa = "all", subtaxa = "Brain", method = "sou")
exp_tree <- NJ(dismat)
exp_tree <- no0br(exp_tree)
var_mat <- varMatInv(objects = tetraExp,phy = exp_tree,taxa = "all",
subtaxa = "Brain")
exp_table <- exptabTE(tetraExp, taxa = "all", subtaxa = "Brain")
exp_one <- aee(exp_table[1,], exp_tree, var_mat)
exp_tree$node.label <- exp_one$est
exp_tree <- root(exp_tree, outgroup = "Chicken_Brain",
resolve.root = TRUE)
plot(exp_tree,show.node.label = TRUE)

}
