% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot.exphy}
\alias{boot.exphy}
\title{Bootstrapping expression phylogeny}
\usage{
boot.exphy(phy = NULL, x, outgroup = NULL, method = c("sou", "pea",
  "spe", "euc", "cos", "jsd", "tani", "jac", "sou_v"), B = 100,
  block = 1, trees = FALSE, quiet = FALSE, rooted = is.rooted(phy),
  jumble = TRUE, mc.cores = 1)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{x}{an exprssion level table: column corresponds to median expression value of all biological samples
within one taxa_subtaxa group; row corresponds to othologous genes}

\item{outgroup}{a vector of mode numeric or character specifying the outgroup}

\item{method}{specifying which distance method to be used
to estimate expression phylogeny in bootstrapping.}

\item{B}{the number of bootstrap replicates.}

\item{block}{the number of columns in x that will be resampled together (see details).}

\item{trees}{a logical specifying whether to return the bootstrapped trees (FALSE by default).}

\item{quiet}{a logical specifying whether to print more information on the screen
while performing bootstrapping(FALSE by default).}

\item{rooted}{if "phy" is a rooted tree, a character of the root node's label when constructing "phy";
if "phy" is unrooted tree, NULL (NULL by default).}

\item{jumble}{a logical value. By default, the rows of x are randomized to avoid artificially
too large bootstrap values associated with very short branches.}

\item{mc.cores}{the number of cores (CPUs) to be used (passed to parallel).}
}
\value{
similar to \code{boot.phylo} in \code{ape}, \code{boot.exphy} returns a numeric vector
which \emph{i}-th element is the number associated to the \emph{i}-th node of \code{phy}.
If trees = TRUE, \code{boot.exphy} returns a list whose first element (named "BP") is like before,
and the second element("trees") is a list with the bootstrapped trees.
}
\description{
bootstrap by resampling gene (gene, transcript, exon, etc..)
}
\examples{
library('ape')
data(tetraExp)
dismat <- expdist(tetraExp, taxa = "all",
                 subtaxa = "Brain",
                 method = "pea")
tr <- root(NJ(dismat), "Chicken_Brain", resolve.root = TRUE)
plot(tr)
exp_table <- exptabTE(tetraExp, taxa = "all", subtaxa = "Brain")
bs <- boot.exphy(phy = tr, x = exp_table, method = "sou",
                 B = 100, outgroup = "Chicken_Brain")
nodelabels(bs)

}
