% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{dist.ced}
\alias{dist.cos}
\alias{dist.euc}
\alias{dist.jsd}
\alias{dist.nbdln}
\alias{dist.pea}
\alias{dist.sou}
\alias{dist.souln}
\alias{distances}
\title{Internal functions for estimating pair-wise expression distances}
\usage{
dist.nbdln(reads.count = NULL, gene_length = NULL, omega = NULL)

dist.souln(reads.count = NULL, gene_length = NULL)

dist.jsd(meanRPKM = NULL, taxon.names = NULL)

dist.pea(meanRPKM = NULL)

dist.euc(meanRPKM = NULL)

dist.cos(meanRPKM = NULL)

dist.sou(meanRPKM = NULL)

dist.ced(meanRPKM = NULL)
}
\description{
Several published methods to estimate pair-wise expression distances
}
\references{
Distance based on negative bio distribution and log normal model
}

