% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{dist.ced}
\alias{dist.cos}
\alias{dist.euc}
\alias{dist.jac}
\alias{dist.jsd}
\alias{dist.nbdln}
\alias{dist.pea}
\alias{dist.sou}
\alias{dist.spe}
\alias{dist.tani}
\alias{dist.u}
\alias{distances}
\title{Internal functions for estimating pair-wise expression distances}
\usage{
dist.nbdln(reads.count = NULL, gene_length = NULL, omega = NULL)

dist.u(reads.count = NULL, gene_length = NULL)

dist.jsd(expMat = NULL)

dist.pea(expMat = NULL)

dist.spe(expMat = NULL)

dist.euc(expMat = NULL)

dist.cos(expMat = NULL)

dist.tani(expMat = NULL)

dist.jac(expMat = NULL)

dist.ced(expMat = NULL)

dist.sou(expMat = NULL)
}
\description{
Several published methods to estimate pair-wise expression distances
}
\references{
Chen H, He X. 2016. The Convergent Cancer Evolution toward a Single Cellular Destination.
Mol Biol Evol 33:4-12

Gu X, Su Z. 2007. Tissue-driven hypothesis of genomic evolution and sequence-expression correlations.
Proc Natl Acad Sci USA 104:2779-2784.

Pereira V, Waxman D, Eyre-Walker A. 2009. A problem with the correlation coefficient as a measure of gene expression divergence.
Genetics 183:1597-1600.

Sudmant PH, Alexis MS, Burge CB. 2015. Meta-analysis of RNA-seq expression data across species, tissues and studies.
Genome Biol 16:287.

Distance based on negative bio distribution and log normal model
}

