% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation.R
\name{estParaWBayesian}
\alias{estParaWBayesian}
\title{Bayesian estimation of expression conservation from previous data set}
\usage{
estParaWBayesian(qgene = NULL, gammaparas = NULL)
}
\arguments{
\item{qgene}{a vector specifying parameter Q estimated from \code{estParaQ}}

\item{gammaparas}{a vector specifying parameters estimated from \code{estParaGamma}}
}
\value{
returns a vector of parameter Ws estimated from previous Q estimation
}
\description{
Bayesian estimation of expression conservation from previous data set
}

