% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation.R
\name{estParaWBayesian}
\alias{estParaWBayesian}
\title{Bayesian estimation of expression conservation from previous data set}
\usage{
estParaWBayesian(Q_gene = NULL, gammaparas = NULL)
}
\arguments{
\item{Q_gene}{a vector specifying parameter Q estimated from \code{estParaQ}}

\item{gammaparas}{a vector specifying parameters estimated from \code{estParaGamma}}
}
\value{
returns a vector of exptations and confidence intervals of parameter Ws estimated from previous Q estimation
}
\description{
Bayesian estimation of expression conservation from previous data set
}
\examples{
data('tetraExp')
species.group <- c("Human", "Chimpanzee", "Bonobo", "Gorilla",
"Macaque", "Mouse", "Opossum", "Platypus")
inv.corr.mat <- corrMatInv(tetraExp, taxa = species.group, subtaxa = "Brain")
brain.exptable <- exptabTE(tetraExp, taxa = species.group, subtaxa = "Brain" ,logrithm = TRUE)
gamma.paras <- estParaGamma(exptable =brain.exptable, corrmatinv =inv.corr.mat)
brain.Q <- estParaQ(brain.exptable, corrmatinv = inv.corr.mat)
brain.post<- estParaWBayesian(brain.Q, gamma.paras)
}
