% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdist.R
\name{expdist}
\alias{expdist}
\title{Expression distance matrix generated from a \code{taxaExp} object}
\usage{
expdist(objects = NULL, taxa = "all", subtaxa = "all",
  rowindex = NULL, method = c("sou", "sou_v", "pea", "spe", "euc",
  "cos", "jsd", "tani", "jac"), logrithm = TRUE)
}
\arguments{
\item{objects}{a vector of objects of class \code{taxonExp} or an object of class \code{taxaExp}}

\item{taxa}{one single character or a vector of characters specifying main taxa selected for
calculating expression distance.
If one single character "all" is given,
all the taxa included in the \code{taxaExp} will be matched and selected ("all" by default).}

\item{subtaxa}{one single character or a vector of characters sepcifying sub taxa selected for
calculating expression distance.
If one singke character "all" is given,
all the subtaxa included in the \code{taxaExp} will be matched and selected ("all" by default).}

\item{rowindex}{a vector of numbers corresponded to indices of selecting rows}

\item{method}{specifying which distance method to be used
to estimate expression phylogeny in bootstrapping.}

\item{logrithm}{a logical specifying whether to apply expression value log2 tranformation (TRUE by default).}
}
\value{
returns an expression distance matrix
}
\description{
Generate an expression distance matrix from an object of \code{taxaExp} class
using a specified distance method
}
\examples{
data(tetraExp)
library('ape')
dismat <- expdist(tetraExp, taxa = "all",
                 subtaxa = "Brain",
                 method = "pea")
tr <- root(NJ(dismat), "Chicken_Brain")
plot(tr)

}
