% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nj.R
\name{NJ}
\alias{NJ}
\title{Neighbor-joining}
\usage{
NJ(X)
}
\arguments{
\item{X}{a distance matrix}
}
\value{
returns an NJ-tree object of class "phylo"
}
\description{
The famous neighbor-joining tree estimation function from Saitou and Nei (1987).
}
\examples{
library('ape')
data(tetraExp)
dismat <- expdist(tetraExp, taxa = "all",
                 subtaxa = "Brain",
                 method = "pea")
tr <- root(NJ(dismat), "Chicken_Brain")
plot(tr)

}
