% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aee.R
\name{no0br}
\alias{no0br}
\title{No negative or zero branch length}
\usage{
no0br(phy, remove.0length = TRUE, root.length = FALSE)
}
\arguments{
\item{phy}{an tree of 'phylo' class}

\item{remove.0length}{a logical specifying whether to replace
negtative branch length with 0 (TRUE by default)}

\item{root.length}{a logical specifying whether to assign a
small value to the branch between root and MRCA of
the ingroup (FALSE by default). If TRUE, the tree should be rooted first}
}
\value{
returns the tree without zero-length or negative-length branch
}
\description{
This function does a small tweak on the tree to
replace negative-length branch with zero-length and/or assign
zero-length branch between root and MRCA of the ingroup a small value
}
