% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tetraexp-data.R
\docType{data}
\name{tetraexp}
\alias{tetraexp}
\title{taxaExp object created from expression data}
\format{a \code{taxaExp} object, contains 53 \code{taxonExp}:
with reads count data and RPKM value on 5636 genes,
raw data and outlier removed data provided.}
\description{
large \code{taxaExp} object created from
six tissues' expression reads count data of nine tetrapod species
}
\examples{
data(tetraexp)
tetraexp.objects
tetraexp.objects[[1]]
}
\references{

}

